Move allocate/memory to memory/allocate
